Avoid use of pixel font sizes in UI
Modify UI CSS to avoid use of 'px' for specifying font sizes, and use
'em' instead. This is useful because sizes specified in pixels may be
wrong (as in, produce different-than-intended results) on even "typical"
high-DPI displays, and in more exotic cases may be outright unreadable.
It seems that using just about any other unit gives more consistent
results across various non-standard scaling/DPI configurations, and
using 'em' should be more responsive to user style overrides.

This is not a complete patch, as it does not attempt to address margins,
padding, and other such attributes which are also currently specified in
pixels, but (for the reasons above) probably should be specified as
something else.