Attempt to fix filter race between style and child updates.
The freeze in 1512 is caused by the group's style property 'filter' connecting
itself to the signal for it's originating object's update. This update causes
a infinate loop as the parent cascades to the filter child which updates any
connected using objects which turn out to contain the self-same filter.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1512