Exact collision detection for the Eraser Tool
This commit is a fundamental reimplementation of the process whereby the
Eraser tool selects the objects to erase.  It introduces a new collision
detection step which ensures that the Cut and Clip modes leave alone the
shapes that weren't touched by the eraser trace. This eliminates many of
the "drive-by damage" bugs that annoyed our users.

The old collision detection step (based on bounding boxes) is still used
but only as the coarse stage of the collision detection algorithm.  Only
the items that pass this coarse test will undergo the second, exact step
which is computationally more expensive. This second stage uses the path
vector of a shape to represent its exact geometric bounds and falls back
to bounding boxes only for items such as raster images. Nonetheless, the
bounding boxes are still transformed in this case: e.g., a rotated image
will have a rotated rectangle as its hitbox.

Furthermore, the collision step correctly accounts for the use of clones
and groups with transformations. This is especially important in the Cut
mode of the tool. This mode is able to unlink clones, but this will only
happen when the eraser path has touched the cloned visual representation
of an item that can be erased in Cut mode (with a boolean operation) and
that has passed the exact collision detection. The new algorithm is able
to analyse the original, underlying shapes by copying their exact bounds
as path vectors (with the necessary transformations), in order to deeply
inspect a cloned group without unlinking the clone.

There are some limitations to this implementation: a text element always
has a rectangular bounding box, although this rectangle will at least be
transformed (the hitbox of a rotated text is a rotated rectangle). Based
on these changes, an exact collision detection with text elements should
not be too hard to add in the future.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3114
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3383