Disable undo if TextEdit widget is under focus
Temporarily disable undo-redo actions when so that the user don't
end up loosing any unsaved data in the widget. Plan is to handle
undo-redo events separately when Inkscape switches to GTKMM4 because
it supports undo-redo operations for each widget individually.

This fix added an event listener that keeps checking for
undo-redo events and if that event matches any of the undo-redo
shortcuts saved by the user, it ignores that call to undo or redo.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/744