Add an action mode to take alignment target from prefs
The actions need to be able to ask the prefs what the last used target
mode was, in order for keyboard shortcuts to do the expected thing.

This also cleans up the use of prefs in the node tool alignment and
while there are no keyboard shortcuts that effect text and nodes the
code has been regularised so there's no surprises for future devs.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3379