Remove depricated preference from prefs skeleton
The iconTheme preference has been depricated and changing the icon theme
via the preferences would clear this value. Each load the skeleton would
put the value back, thinking it way missing. Overriding the set value.

This fix also includes setting the theme to "custom" if any theme prefs
are changed. This turned out not to be the problem, but it's a good UX
fix.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3396