A keyboard shortcut observer for an action
This commit adds a simple utility class `ActionAccel` which stores the
keyboard shortcuts associated to an action and provides an easy way to
check if a given key event triggers one of these shortcuts. By keeping
track of changes to the keybindings, each instance automatically stays
up to date with the keyboard shortcuts for the observed action.

Additionally, several string copies in the keyboard shortcuts code are
prevented through taking range-for loop variables by reference.