Private redo for the in-progress path of Pen Tool
The Pen/Bézier Tool reacts to the (customizable) keyboard shortcuts
associated to the "undo" and "redo" actions and performs its own
internal equivalent of undo and redo operations on the path that is
being drawn. As before, the internal "undo" is simply the removal of
the last added point from the in-progress path.

This commit implements a new internal "redo" functionality which
restores points previously removed by the internal undo.

Closes https://gitlab.com/inkscape/inbox/-/issues/2935
Closes https://gitlab.com/inkscape/inbox/-/issues/6755