Ensure actual focus for a newly opened dialog
Ensure that the child widgets of a freshly opened dialog can be
immediately traversed with the Tab key. Currently, the contents of each
dialog are repackaged into a Gtk::ScrolledWindow for the purpose of
better controlling the scrolling behaviour. This ScrolledWindow must not
steal the focus from the widgets that the user cares about.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3482