Fix padding for text shape-inside creating triangles
The padding shape is trying to do an inset of a starting shape-inside
but it did not take account of what the livarot offset code was doing
producing all sorts of weird artifacts as the "inset" is not accounted
for properly.

This also fixes an introduced problem with exclusion margin handles not
moving with the mouse. Inverted sign error.

Fixes https://gitlab.com/inkscape/inbox/-/issues/6909