Add support for DrawingItem snapshotting
Implement a snapshotting mechanism to allow Drawing and DrawingItems
to be temporarily bitwise-frozen while rendering is taking place. While
frozen, all modifications are temporarily diverted into a FuncLog on a
far-away cacheline to be replayed later. This provides a step towards
thread-safe rendering in a minimally invasive way by taking advantage
of the existing object-tree/display-tree split.

All smart pointers in the object tree are converted from unique_ptr to
DrawingItemPtr (a unique_ptr with a custom deleter) to add the necessary
support for deferred deletion. A slight modification to how DrawingText
references FontInstance was required to make its destructor safe to run.