Allow for rasterization of FEs outside page one
When an item with filter effects is rasterized, its bounding box is only
clipped to the nominal bounding box of the document when the document is
in "legacy" mode (single visible page = zero SPPages). When the document
has actual SPPages in it, the item's bounding box is clipped to the page
that is being written. This enables rasterized filter effects to show up
on pages of PDF output after the first one.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3214