Correctly fit page to selection with y-axis up
When the y-axis points up, the reference point for shifting the drawing
items and scrolling the canvas view should be the max() rather than the
min() of the y-interval of the desired bounding box.  Clean up the code
and remove the remnants of margins functionality (soon to be replaced).

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3609

Co-authored-by: Martin Owens <doctormo@gmail.com>