PDF output: round pt drawing size up to an integer
PDF and PostScript documents created by Cairo specify page dimensions
as integers (in PostScript points), so the pages in export output end
up slightly larger than what Inkscape requests. This change gives the
user the option to either keep the relative size of the drawing (with
respect to the page) or to keep the absolute sizes. The new option is
selected from the PDF export dialog.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3569