Make SPPattern create persistent display tree
Change the rendering strategy of SPPattern to the same as SPHatch.
That is, instead of creating and tearing down display tree on every
render, create persistent display tree instead.

* This required significant modifications to SPPattern in order to
  reliably keep the display tree up-to-date with the object tree.

  Most of the complications come from the fact patterns can
  reference other patterns, forming href chains.

* The display key system also had to be reworked to avoid key
  collisions.

* Fixed a bug in SPStyle resulting in missing updates when a paint
  server becomes invalid.

* Removed a no-longer necessary hack from the attribute dialog.

On its own, this commit may result in higher memory usage due to
the increased caching. This is fixed by the next commit.