Prevent pattern skewing in the tiled manipulator
Restore old transform behaviour whereby only non-uniform scale and
translation can be applied to a pattern. Fix a display glitch when
the y-axis points up. Prevent numerical instability resulting from
the accumulation of floating point errors by caching the transform
at the beginning of a drag and using it on every update. Limit the
maximum number of pattern repetitions in an item to 1e6 to prevent
Cairo from glitching out and drawing a black fill.

Co-authored-by: Martin Owens <doctormo@gmail.com>