Custom GUI extensions supress stderr warnings
When an extension issues stderr logs, but returns a successful
output we show a warning to the user in case there's something
they should be aware of.

For custom GUI extensions (extensions manager, import web, etc)
there are many python libraries which log to stderr about all sorts
of useless things which can not be supressed. Yet a GUI application
has all the time to tell the user anything important in it's own
interface, should it need to.