Rewrite pattern rendering, fix memory spike
This commit rewrites DrawingPattern and related code to reduce
memory usage when rendering patterns.

The main change is to only draw the visible area of patterns.
This involves tracking this information through the DrawingItem
hierarchy and NRStyle so it actually reaches DrawingPattern.
Visibility is determined both by the area being rendered and the
size of the object with the pattern. Then, using a periodically-
wrapped cache, only the part of the pattern tile that is visible
and hasn't been drawn before is actually drawn. NRStyle was
modified to always communicate with DrawingPattern, rather than
cache its pattern when first generated, because more pattern could
need to be drawn at any time.

Caveats: Pattern tiles continue to live outside the cache system,
invisible content is never discarded, and oversampling is disabled
to avoid the exponential growth of tile sizes for nested patterns.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1018