Fixed font-family combobox rendering issue
The font-family combobox is now gradually and recursively populated, after the first popup callback, with groups of entries with size = FONT_FAMILIES_GROUP_SIZE, so to not overload the total rendering of the list.
The patch fixes an issue of GTK-3 combobox, which forces the user to popup twice the combobox when it's initially displayed, in order to see the complete list of font-families including samples and scrollbars.