Symbols dialog: place symbol at the drop location
When a symbol is dragged from the Symbols dialog, it will be inserted
into the document at the mouse cursor position it is dropped at, with
the anchor point being the part of the symbol where the drag started.
Additionally, we fix a memory leak from recreating tree model columns
(the `return new` antipattern).