Fix a use-after-free after exporting with Cairo
Inkscape will no longer crash after completing a PDF/PS export when the
exported document contains a path with a gradient as fill or stroke and
a marker referencing that fill or stroke. To prevent the segfault, make
the replacement of pointers to paint servers temporary and restore them
after the marker has been rendered.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3456