Measure from the hovered item's desktop affine
When measuring an item, we use the i2dt too get an affine
which gives us the pixel values. Converting to the display
unit is then done afterwards.

Also fixes a problem of consistancy with the measure toolbar
when the measure unit preference isn't set. It would be
"mm" in code (default unit) but "cm" in toolbar (first item).

Fixes https://gitlab.com/inkscape/inkscape/-/issues/1367