Add action "paste-on-page" to allow pasting on any page.
This will apply the copied page's offset to pasted items, allowing
items from one page to be pasted into the same position on a diffent page.

The style menu has been tweaked to fold the extra paste options into it's own
sub menu.

The new and delete actions are amended to focus the window on them when used
from the toolbar.