Detect and add in missing letters from bad Unicode
Rarely a PDF will have two letters right next to each other be detected
as a single unicode. When we convert them to utf8 they split into two
but we only add the first letter, causing letters to be missing from the
resulting svg.

This uses tspans to allow us to add the extra letters without disturbing
the position of the following glpyhs.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3771