Add option to copy style verbatim, and also class
Add an option to control how an object's style is copied and pasted
(both normally and with "paste style"). One option (and the default) is
the historic behavior, which applies the computed style to the object on
the clipboard. The other option copies the "style" attribute — and also,
newly, the "class" attribute — verbatim. The latter may result in the
pasted object or style differing from what was copied, but may be the
desired behavior for documents making use of CSS for styling.

Note that this may result in unusual behavior when copying and pasting
between different instances which have the option set to different
values.