Fix moving two items from different layers
If one of the items in the selection is the last items of the
target layer, then we ended up with premature return. with this
fix we are able to not attempt to move the problematic object
and just reselect it at the end.

Cleanup to remove some more undo boolean patterns.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3778