Fix overlapping infobox in measurement tool
This changes fix the overlapping of infobox and cursor point of
measurement tool when cursor hover over any shapes. It also fixes
difference in x-shift of infobox from cursor point on different y-axis
direction. I shifted the infobox above as well as to the right of  cursor
point to overcome the overlapping.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3756