Grid Refactor
Currently a lot of the work is manually done by
`src/display/control/canvas-grid.cpp` - almost everything from
construction to rendering.
`src/object/sp-grid.cpp` does not exist in the project, so the goal is to
integrate the SPGrid object and remove the canvas-grid.* source files.

There are several steps to breaking up canvas-grid:
[*] Extraction of XML repr related code to new SPGrid
[*] Have CanvasItemGridsAxonom subclass CanvasItemGrid
[*] Proper updating and rendering of CanvasItemGrids
[*] Extract Gtk Widget related code to DocumentProperties
[*] Move snapping to their own classes