Remove XML::NodeEventVector
XML::NodeEventVector was a C-style way of emulating virtual functions
which doesn't need to exist anymore. This paired with
XML::Node::addListener was used to GC-allocate observers while not
having to care about their lifetimes.

Removal was accomplished by inheriting from XML::NodeObserver in classes
which handled the code for adding and removing observers in most cases.
In some cases (see NodeData in src/widgets/sp-xmlview-tree.cpp), a
pointer to an XML::NodeObserver is used to provide a more dynamic
approach.

Additional minor refactoring is also performed, including the moving of
default initialization values from ctors to header files.