Change CompositeNodeObserver::ObserverRecordList into a std::vector
With the removal of XML::NodeEventVector and XML::Node::addListener(),
XML::Observers in an ObserverRecordList are no longer GC-allocated,
hence they do not need to be scanned.

This allows us to switch from std::list to std::vector which is almost
always more efficient. The vector is also made unscanned.

GC::Alloc was given an extra template parameter to make this possible.