Fix cairo pdf import page selection
The page selection must be a valid value in order to be parsed
and the first of the list to be used in the cairo importer.

Before this was always set to "all" and so only the first page
would be imported.

Added some other cleanups

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3885