Fix PDF export of markers on cloned items
Ensure that markers using context-{fill,stroke} can find the appropriate
style context in the scenario when the object using them has been cloned
and the <use> element doesn't set fill/stroke explicitly.

In order to fix the issue, we ensure that the styles of the use item and
its child are merged, in a manner similar to what is done when unlinking
a clone. This operation and the temporary rewriting of styles needed for
the handling of context-fill and context-stroke are performed by an RAII
class which manages the rewriting of paints, restoring the old values in
its destructor.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3911