Repair layer and object blend mode list
Previous fix created seperators in blend mode list which was
not taken account in the layers and objects interface.

This uses the seperators to decide how to lay out the blend
modes, keeping the original method as much as possible.

A better solution would be to use glade, which we can do in
the future.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3926