Support for OKLab/OKLch and a new colour picker
Add support for OKLab and OKLch, perceptual colour spaces selected for
the CSS Color Module Level 4 draft recommendation. Create a new colour
picker based on OKLch.  The new picker mode is called OKHSL and can be
selected from the dropdown menu. The icon used in the menu was created
by alezzacreative.

Note that the OKHSL colour space is not standardised, but the picker's
appearance can easily be tweaked if a standard is ever issued.

A few light refactors of ColorScales are needed, mainly replacing float
variables with doubles, necessary to achieve a reasonable precision of
colour conversions.

Co-authored-by: alezzacreative <alezzacreative@gmail.com>

Closes https://gitlab.com/inkscape/inbox/-/issues/7733