Replace royalist imagery by peaceful one
This replaces the coat of arms with a fleur-de-lys used as the icon for
the symbols tab by the peace symbol. This imagery is used as far as I
know in two ways, in Quebec as a nationalist imagery, and in France as
the very-far-right symbol of the royalists, used by several dangerous
organizations like Action Française. I think none of these ideologies
should be represented, so I replaced it with a globaly positively
regarded one.

I spend lots of time removing stickers and posters with the symbol
covering the walls of my city, I didn't expect I'd have to remove it
from a software I use very regularly. ⭩⭩⭩