Add syntax coloring to the XML Dialog
Co-authored-by: Rafael Siejakowski <rs@rs-math.net>

Use GTKSourceView4 to implement syntax highlighting in the XML Dialog
for CSS editing (both the contents of the <style> tag and inline styles)
as well as for the SVG path `d` attribute and the `points` attribute of
poly{line|gon} elements. Consistently-styled XML coloring is also used
in the XML tree shown in the dialog, by recognizing XML elements and
attributes.

Several additional UI improvements are made to the XML Dialog:

* When editing the `d` attribute, a number-rounding button is available.
* The `d` attribute and inline CSS are automatically reformatted for
  better readability.
* The attribute editing popover does not close on Enter press in these
  modes. Its resizing behavior and timings are improved.
* The superfluous `svg:` namespace prefix is not shown in the XML tree.
  When the user creates a new node via the dialog, the prefix is
  accepted but not required.
* A validation step is added for new element nodes in SVG namespace; any
  invalid elements are highlighted as errors.

The user can select the coloring theme and the monospaced font for use
in the XML Dialog. Two custom made, Inkscape-specific themes are added,
one for dark and one for light UI mode. The GTKSourceView4 library is
added as a dependency in the msys2 and Snap packaging environments.

Closes https://gitlab.com/inkscape/inbox/-/issues/4263
Closes https://gitlab.com/inkscape/inbox/-/issues/4264