Fix windows build after pango upgrade (1.50-11 to 1.50-12)
Pango now requires dwrite, which can be enabled in meson. The
logic for enabling it was not written in autotools.

Switch to meson avoided on Linux as our appimage CI/CD is based on
a Ubuntu 20.04 docker base, but cairo requires meson >= 0.56.0.
Switch isn't required yet as dwrite is windows specific backend.

This stops disabling trace and script-interpreter, which doesn't seem
possible with the meson scripts. Prefix/libdir were also unset, see
https://gitlab.com/inkscape/inkscape/-/issues/3862. It installs cairo
to the system (which is dangerous!). However, none of these options
were previously used in Windows builds. OUTPUT_STRIP_TRAILING_WHITESPACE
was missing from the executable_process call that defined AUTOGEN_PATH,
so autogen.sh was called without any arguments.