Fix crash deleting row, Selectors dialog on Windows
_deletion doesn't seem necessary (only used in one
place). I've removed that use, replacing it with a
defensive check to avoid crashing, though we might
end up selecting a different cell now. Now _deletion
is only true during the call to _onPropDelete, and is
used to stop readStyleElement from running and deleting
the tree view, which is still used by a caller of
_onPropDelete (gtk_tree_view_multipress_gesture_pressed).

Fix https://gitlab.com/inkscape/inkscape/-/issues/3986