Add document resources dialog
This is a new dialog that adds insight into document's assets,
some of which may not be obvious, as they are not necessarily rendered.
Many of the resources reside in a <defs> element.
This dialog can present an overview of a document resources, as well as
a drill down to examine individual resource types.
It presents colors, swatches, gradients, patterns, symbols, markers,
metadata, and more.
The dialog also offers some limited editing capabilities (labels,
deletion) plus image and colors extraction.