Implement smooth auto-scrolling
Implement support for having the canvas smoothly scroll when an object
reaches the edge. Such objects include knots, control points, dragged
objects and the selection rectangle. The animation begins when a tool
calls enable_autoscroll() while the mouse is dragged beyond the canvas
edge, and stops when the mouse is released. (Having tools cancel the
animation would make more sense, but this proved too unreliable due to
missed release events.) The function previously used for auto-scroll,
scroll_to_point(), remains but is now exclusively used for warping the
view instantly to an object.