Fix Crash in cropToObjects
When a clone or text is linked to an object that will be deleted the process doesn't
retain the original object. It sometimes successfully unlinks depending on the order
of processing, but can quite often crash the Export dialog on multiple pages to pdf.

To fix this we retain any linked objects.