Further fix CanvasItem visibility system
Document the invariants that the canvasitem visibility system relies on,
and insert a _need_update = false assignment to fix a case where they
could be broken. Specifically, if an item was marked as needing an update
while it was invisible and not needing an update, then upon becoming
visible again, it would never be updated again until its parent was
updated.

Additionally, streamline Handle::setVisible() and fix a code-rot issue in
Canvas::paint_widget().