Fix duplicate extension ids and bad errors on null
The extensions system couldn't cope with two extensions using
the same id, even though this is very likely in the case where
a user uses a newer version of an extension in their config folder.

We now deregister the previous extension, which is most likely the
system extension. It's last to register wins.

Also fixes an error where no-action would cause an error that the
contents couldn't be parsed. This is a regression.