Fix sort order of recent files in startup dialog
We want to make sure the user can find the latest files
they have modified when using the start screen. So we
sort them by their last modified date.

Other improvements are recording when inkscape crashes
the crash file in the RecentManager which can be detected
by the startup screen and labeled.

This allows users to open crash files and keep working
as if they were the original file.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4081