Fix opening different inkscape versions at the same time
This patch adds the revision hash to the app id, forcing Inkscape
to allow different versions of Inkscape to be installed on the
same computer without getting in it's own way.

This does change the default DBus address, and it might make it more
unpredictable. So this should be tested to see what impacts it has.