Don't set undo if nothing has changed in shape builder
Previously we were depending on the undo system to know
if nothing had changed in the document. But now we are
more explicit.

Also allow the confirm button to function even if nothing
has changed with the setup. The user is being explicit.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3906