Remove pdf/ps specific size, bleed and exportId options
These out dated size, bleed and export options are all covered
by various other bits of functionality and should not be duplicated
in both the new export dialog and pages tool as well as the large
amount of duplication and inconsistancy in these options between
each of the pdf and postscript extensions themselves.

For the command line options, we re-use the svg pathway to reduce the
amount of duplication and achieve the same export options across all
of the vector outputs.

An information section is added to the preferences for each of the
extensions explaining how to set the removed properties. An upgrade
to the widget-image object was needed to support inkscape-icon names.