Remove async preview rendering and use a single drawing object
To improve load times for documents with multiple documents we
create a single drawing document for each of the export types
and share it between all the previews.

Rendering then uses this same construct for each different preview.

This allows the GUI to update while previews are updating and won't
get in the way of the user editing on the document.

Fix: Crash in page-delete caused by sorted comparison of removed page
Fix: Lack of updating because of new overlay widget parents