Fix scroll issues with node selection
Currently the node tool will selectively steal the scroll inputs
in order to select nodes, if and only if the user is hovering over
a node point.

This small target, plus the always on default configuration leads
to a UX problem where users expecting to zoom in and out will instead
select more nodes, disrupting their flow.

This changes makes the node selection feature configurable. But does
not change the defaults.