Reinstate missing drawing update before pick
In 19702392, a drawing update was removed from pick_current_item()
because it would later become unsafe to call while rendering is taking
place. But if it isn't, then this call is not only safe but necessary,
so this commit adds it back (with the requisite check).

Fixes an unreported 'Invalid state when picking' bug resulting in an
inability to drag an object that had just been dragged.