Fix the repeating extension preference popups in batch export
When exporting multiple formats the batch exporter would require
the user to confirm the preferences for each run. This limited
how easily it was to run a batch export while doing other work.

A new button is added to the single and batch export next to the
preferences dropdown so the user can choose to edit the preferences
of the export format and they don't need any in-progress popups.

The seperation of preferences for each batch format row requires
a much larger refactoring of the extensions system and has been
split out.