Allow values with commas in object-set-attribute
The CLI action object-set-attribute will now be able to set an attribute
to a value containing a comma. Although a single comma is still needed to
separate the attribute name from the desired value, any subsequent commas
will be interpreted as a part of the value.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3685