Fix infinite recursion in filter bbox calculations
Fix an infinite loop occurring when the same filter is applied to a
group and its parent group. The fix is to replace the existing call
to `SPGroup::bbox()` in filter code by a call to the newly implemented
function `SPGroup::visualBounds()`, capable of ignoring filters.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/3797