Remove signal destruction for document uri and widgets
The XML dialog was over egging it's signals connections by disconnecting
them on un_realize. This is not a safe way of managing signals since the
widgets get unrealized when being dragged into floating dialogs.

The C based signals and the use of g_objects is a 20 year old codebase
which is not in keeping with how we do C++ today, so the signals are
replaced with Inkscape::auto_connect to make better behaved signals.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4187