Feature: PDF Internal Linking
We have supported linking to external websites using urls in pdf for
a long time. But have not supported linking to other pages or items.

This feature adds tracking of connected objects to the sp-anchor so we
know when a hyperlink is linked to an local object instead of a url.

The fragment syntax also works if you load the svg in a web browser.

Then the cairo rendering stack is modified to support linking to both
objects and pages using destination linking. Pure page linking was
attempted but failed to work correctly.