Save attributes containing \n to XML correctly
The XML specification is vague about what happens when an xml value has
a line feed in it when writing. What we know is that it's not supposed
to contain an actual line feed character and any that it reads are
converted to spaces.

Our xml parser handles this correctly. But our writing process does not
blindly writing line feeds right in the middle of the attribute.

This fix replaces line feeds with the xml recommended &#10; which is read
back in as a line feed correctly and parsed by other xml programs as intended.

Also added is a way to use CTRL+Return to add line feeds in the xml editor
mostly for testing and advanaced users.