Fix recently used files on export and default filename
The export batch functionality was populating the recently used files
with a lot of filenames, rendering the recently used menu list pointless.

We make it so that the single export adds a recent file, but batch does not.

We also repair the single export loading order, which was attempting document
updates before the document was loaded into the dialog.