Ensure text toolbar style is always being used
The text toolbar style should come from two different places depending on the
option that is set. The option 'usecurrent' which is our default, takes the
text style from the desktop global style, but all the toolbars are wired up
to modify the tools own internal style, which is never used when in this mode.

To fix this we reconfigure both reading and writing of the styles such that
they modify the correct preference. We reduce the amount of duplicate code
and make the toolbar consistant between the different styles.