Allow the user to control when layer and page changes
Changing the selected objects will always change the page and layer, but there
are lots of contexts where we don't want this to happen and it's disruptive to
the user's workflow.

We add a boolean control for each functionality into the select object.

We then add a tool-based setting, allowing each tool to control the preference.
The preference is shown for the select and node tools, all other tools are set
to false so using the rectangle tool or text tool won't ever change the layer
or page.

Select is default to true for both and node is default to false, but is changable.