Add groups to hold compounded clipping groups
Some PDF files contain groups with multiple clipping regions. Our new PDF code
tries to reduce the number of groups it makes for each state change, but in this
case we need to actually apply the clipping path and that means setting up groups.

We also want the clipHistory maanger to error out if we try and over-write clipping
paths. This is an indication of data loss.

Test changes have been confirmed as ok.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4245