Combine glyphs of the same text span in text-to-path
We want to preserve the text objects as much as possible when converting
from text to a path. The paint order depends on objects which are combined
in order the strokes don't overlap their siblings.

In this fix we loop through each set of glyphs for each of the parents and
make a path for each of them. If we end up with more than one path, then we
make a group to contain them and distribute the styles and transforms.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4256
Fixes https://gitlab.com/inkscape/inkscape/-/issues/4261