Ungrab handles on destruction
Because canvas item destruction can be deferred, the ungrabbing
can happen after the creation and grabbing of a new handle. The
grabbing of a new handle is blocked by the existing handle and
causes the new handle to be 'dropped' in the middle of it's work
leaving the document in a bad state.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4239