Fix picking order in outline mode
Respect outline mode in getItemFromListAtPointBottom() and friends.
This requires making these functions canvas-aware, so they can check
whether the query point is in the outline region or not. Before the
regression, it just used whether the last mouse position over the canvas
was in the outline region. This gave the appearance of working, because
the two points are usually the same.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4271