Request gtk size updates on_map instead of create
For dialogs, we have a better experience on load if we request a size
allotment on_map instead of on creation. This means the widget will
have a parent and the allotment will make sense.

We also want to do the update on the dialog itself, so child widgets can
be alotted the right sizes. This helps with the object/layer dialog and
it's vanishing entries.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/2373
Fixes https://gitlab.com/inkscape/inkscape/-/issues/3785