Remove Pixmap Directory
This refactoring of the share/pixmap directory moves away the elements
which were still in use, specifically the branding icons from the weclome
screen which are moved to the share/screens directory with their siblings
and the two io/resources uses for color-item and export-dialog which are
moved to share/ui/resources finally the inkscape icon in various sizes is
moved to the share/branding directory.

All other graphics are not in use and are removed.