Resize align dialog icons when requested
We needed to allow the user to increase size of the icons in the align
and distribute dialog as a matter of accessibility and user experience.

This patch is a more raw form of the full patch suitable for the 1.3 release and
contains no string changes. It uses the toolbox icon size as the source of preference

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4165